*: move pkg/flags to pkg/environment
