+++ components/playground/main.go
@@ -568,7 +568,7 @@ func main() {
 		f := func() {
 			defer func() {
 				if r := recover(); r != nil {
-					if localdata.DebugMode {
+					if environment.DebugMode {
 						log.Debugf("Recovered in telemetry report: %v", r)
 					}
 				}
@@ -593,7 +593,7 @@ func main() {
 			tele := telemetry.NewTelemetry()
 			ctx, cancel := context.WithTimeout(context.Background(), time.Second*2)
 			err := tele.Report(ctx, teleReport)
-			if localdata.DebugMode {
+			if environment.DebugMode {
 				if err != nil {
 					log.Infof("report failed: %v", err)
 				}
